move out the utility functions from Cycle.scala to CycleUtils.scala
